--- src/app/components/shared/checkbox/checkbox.less
@@ -1,46 +1,0 @@
-@import "./../../../../../node_modules/nav-frontend-core/less/_variabler";
-
-.checkbox {
-	border: 1px solid #C6C2BF;
-	border-radius: 8px;
-	padding: 15px;
-	display: flex;
-
-	&__success {
-		background-color: #cde7d8;
-		border: none;
-	}
-
-	&__info {
-		background-color: #ffe9cc;
-		border: none;
-	}
-
-	&__normal {
-		border-radius: 8px;
-	}
-}
-
-.checkbox__success > .checkboks:checked + .skjemaelement__label:before {
-	border-color: @navGronn;
-	background-color: @navGronn;
-}
-
-.checkbox__normal > .checkboks:checked + .skjemaelement__label:before {
-	border-color: @navBla;
-	background-color: @navBla;
-}
-
-.checkbox > .checkboks + .skjemaelement__label:before {
-	border-radius: 12px;
-}
-
-.checkbox > .checkboks + .skjemaelement__label:before {
-	width: 24px;
-	height: 24px;
-	top: 0;
-	bottom: 0;
-	left: -6px;
-	margin-top: auto;
-	margin-bottom: auto;
-}
